# docs/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: Kanbien Platform API
  version: 0.2.0
  description: >
    OpenAPI specification for the platform API. This documents the platform
    health route, root-user authentication, root-role authorization
    administration, and the protected rootUsers feature.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/root-auth/login/password:
    post:
      tags:
        - Root Auth
      summary: Start root-user login with email and password
      operationId: loginRootUserWithPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RootAuthPasswordLoginBody"
      responses:
        "200":
          description: Password accepted and SSH challenge issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootAuthChallengeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-auth/login/ssh:
    post:
      tags:
        - Root Auth
      summary: Complete root-user login with SSH proof
      operationId: completeRootUserSshChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RootAuthSshLoginBody"
      responses:
        "200":
          description: Root user authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootAuthSessionSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/root-auth/principals:
    post:
      tags:
        - Root Auth
      summary: Create an auth principal for an existing root user
      operationId: createRootUserAuthPrincipal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRootAuthPrincipalBody"
      responses:
        "201":
          description: Auth principal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRootAuthPrincipalResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-auth/password/change:
    post:
      tags:
        - Root Auth
      summary: Change the authenticated root user's password
      operationId: changeRootUserPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeRootAuthPasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-auth/ssh-keys:
    post:
      tags:
        - Root Auth
      summary: Register an SSH public key
      operationId: addRootUserSshPublicKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRootAuthSshKeyBody"
      responses:
        "201":
          description: SSH key stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootAuthSshKeySummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags:
        - Root Auth
      summary: List SSH public keys for the authenticated root user
      operationId: listRootUserSshPublicKeys
      security:
        - bearerAuth: []
      responses:
        "200":
          description: SSH public keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootAuthSshKeyListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-auth/ssh-keys/{keyId}:
    delete:
      tags:
        - Root Auth
      summary: Revoke an SSH public key
      operationId: revokeRootUserSshPublicKey
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthSshKeyId"
      responses:
        "200":
          description: SSH key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-auth/sessions:
    get:
      tags:
        - Root Auth
      summary: List sessions for the authenticated root user
      operationId: listRootUserSessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootAuthSessionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-auth/sessions/{sessionId}/revoke:
    post:
      tags:
        - Root Auth
      summary: Revoke one of the authenticated root user's sessions
      operationId: revokeRootUserSession
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthSessionId"
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-auth/logout:
    post:
      tags:
        - Root Auth
      summary: Log out the current root-user session
      operationId: logoutRootUserSession
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
...
    AuthSshKeyId:
      name: keyId
      in: path
      required: true
      schema:
        type: string
    AuthSessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed or was not supplied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated caller is not allowed to complete the action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict with existing or current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Request was throttled or temporarily locked due to too many attempts
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    RootAuthPasswordLoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    RootAuthChallengeResponse:
      type: object
      properties:
        status:
          type: string
          enum: [SSH_CHALLENGE_REQUIRED]
        challengeId:
          type: string
        challengeText:
          type: string
        availableSshKeys:
          type: array
          items:
            $ref: "#/components/schemas/RootAuthLoginSshKeyOption"
      required: [status, challengeId, challengeText, availableSshKeys]
    RootAuthSshLoginBody:
      type: object
      properties:
        challengeId:
          type: string
        signature:
          type: string
        publicKeyFingerprint:
          type: string
      required: [challengeId, signature, publicKeyFingerprint]
    RootAuthSessionSummary:
      type: object
      properties:
        status:
          type: string
          enum: [AUTHENTICATED]
        sessionId:
          type: string
        authPrincipalId:
          type: string
        rootUserId:
          type: string
          format: uuid
        authenticatedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
      required: [sessionId, rootUserId, authenticatedAt, expiresAt]
    RootAuthLoginSshKeyOption:
      type: object
      properties:
        keyId:
          type: string
        label:
          type: string
        fingerprint:
          type: string
      required: [keyId, label, fingerprint]
...
    RootAuthSessionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RootAuthSessionSummary"
      required: [items]
    CreateRootAuthPrincipalBody:
      type: object
      properties:
        rootUserId:
          type: string
          format: uuid
        loginEmail:
          type: string
          format: email
        password:
          type: string
      required: [rootUserId, loginEmail, password]
    CreateRootAuthPrincipalResponse:
      type: object
      properties:
        authPrincipalId:
          type: string
        rootUserId:
          type: string
          format: uuid
        loginEmail:
          type: string
          format: email
      required: [authPrincipalId, rootUserId, loginEmail]
    ChangeRootAuthPasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
    AddRootAuthSshKeyBody:
      type: object
      properties:
        label:
          type: string
        publicKey:
          type: string
      required: [label, publicKey]
    RootAuthSshKeySummary:
      type: object
      properties:
        keyId:
          type: string
        label:
          type: string
        algorithm:
          type: string
        fingerprint:
          type: string
        status:
          type: string
          enum: [active, revoked]
        createdAt:
          type: string
          format: date-time
        revokedAt:
          type: [string, "null"]
          format: date-time
      required:
        - keyId
        - label
        - algorithm
        - fingerprint
        - status
        - createdAt
        - revokedAt
    RootAuthSshKeyListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RootAuthSshKeySummary"
      required: [items]
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ErrorDetails:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          $ref: "#/components/schemas/ErrorDetails"
      required: [code, message]